Mini-Project-2/visual_api_designer/requirements.txt
# Django and Core Dependencies
Django>=4.2.0,<5.0
djangorestframework>=3.14.0

# Database and Environment
dj-database-url>=2.0.0
python-decouple>=3.6
psycopg2-binary>=2.9.0  # For PostgreSQL
mysqlclient>=2.1.0      # For MySQL (uncomment if using MySQL)

# Production and Deployment
whitenoise>=6.0.0       # Static files serving
gunicorn>=21.0.0        # WSGI server for production

# CORS for React frontend
django-cors-headers>=4.0.0

# Development Dependencies (optional)
django-extensions>=3.2.0
python-dotenv>=1.0.0

# Optional: For advanced features
Pillow>=10.0.0          # If you need image handling
celery>=5.3.0           # For background tasks (optional)
redis>=4.5.0            # For caching/sessions (optional)